rename exchangeTransactions.js to Exchnage.js
