+++ src/augmintjs/Exchange.Matching.test.js
@@ -1,5 +1,5 @@
 const expect = require("chai").expect;
-const exchange = new (require("./exchangeTransactions"))();
+const exchange = new (require("./Exchange.js"))();
 const BigNumber = require("bignumber.js");
 const { cost } = require("./gas.js");
